Man page says this is an int not a size_t.
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/classlib/trunk@498823 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/luni/src/main/native/luni/unix/OSFileSystemLinux32.c
@@ -234,6 +234,6 @@ JNIEXPORT jlong JNICALL Java_org_apache_harmony_luni_platform_OSFileSystem_trans
   if(hysocketP == NULL)
     return -1;
   socket = hysocketP->sock;
-  return sendfile(socket,(size_t)fd,(off_t *)&offset,(size_t)count);	
+  return sendfile(socket,(int)fd,(off_t *)&offset,(size_t)count);	
 }
